Remove redundant utilities, now that they're only used in one place.
